project/requirements.txt
Django==1.8.2
django-appconf==1.0.1
django-compressor==1.5
django-extensions==1.5.5
django-filter==0.10.0
djangorestframework==3.1.3
djangorestframework-bulk==0.2.1
docutils==0.12
drf-extensions==0.2.7
drf-nested-routers==0.9.0
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
oauth==1.0.1
optional-django==0.1.0
Pygments==2.0.2
pytz==2015.4
six==1.9.0
snowballstemmer==1.2.0
Sphinx==1.3.1
sphinx-bootstrap-theme==0.4.7
sphinx-rtd-theme==0.1.8
wheel==0.24.0